still working on this
I setup a new database of trump tweets to pull from instead of the previous trolls data set. This one seems to pull in through the console AND print correctly in the PrestoAPI website, but I still can't seem to print the data to my page. Feeling like a pooey developer.